Create and reuse base class
