+++ nameko_sqlalchemy/__init__.py
@@ -1,4 +1,3 @@
 from nameko_sqlalchemy.database_session import (  # noqa: F401
-    DatabaseSession, DB_URIS_KEY, Session
+    DatabaseSession, DatabaseSessionScope, DB_URIS_KEY, Session
 )
-from nameko_sqlalchemy.session_scope import DatabaseSessionScope  # noqa: F401
