+++ nameko_sqlalchemy/database_session.py
@@ -1,3 +1,4 @@
+from contextlib import contextmanager
 from weakref import WeakKeyDictionary
 
 from nameko.extensions import DependencyProvider
@@ -7,10 +8,10 @@ from sqlalchemy.orm import sessionmaker
 DB_URIS_KEY = 'DB_URIS'
 
 
-class DatabaseSession(DependencyProvider):
+class BaseDatabaseSession(DependencyProvider):
+
     def __init__(self, declarative_base):
         self.declarative_base = declarative_base
-        self.sessions = WeakKeyDictionary()
 
     def setup(self):
         service_name = self.container.service_name
@@ -23,15 +24,22 @@ class DatabaseSession(DependencyProvider):
             'declarative_base_name': decl_base_name,
         })
         self.engine = create_engine(self.db_uri)
+        self.Session = sessionmaker(bind=self.engine)
 
     def stop(self):
         self.engine.dispose()
         del self.engine
 
+
+class DatabaseSession(BaseDatabaseSession):
+
+    def __init__(self, declarative_base):
+        self.sessions = WeakKeyDictionary()
+        super(DatabaseSession, self).__init__(declarative_base)
+
     def get_dependency(self, worker_ctx):
 
-        session_cls = sessionmaker(bind=self.engine)
-        session = session_cls()
+        session = self.Session()
 
         self.sessions[worker_ctx] = session
         return session
@@ -39,6 +47,26 @@ class DatabaseSession(DependencyProvider):
     def worker_teardown(self, worker_ctx):
         session = self.sessions.pop(worker_ctx)
         session.close()
+
+
+class DatabaseSessionScope(BaseDatabaseSession):
+
+    def get_dependency(self, worker_ctx):
+
+        @contextmanager
+        def get_session_scope():
+            session = self.Session()
+            try:
+                yield session
+            except:
+                session.rollback()
+            else:
+                session.commit()
+            finally:
+                session.close()
+
+        return get_session_scope
+
 
 # backwards compat
 Session = DatabaseSession
